fix: 채팅 기능 읽어 오기 등 완성

+++ src/main/java/dogveloper/vojoge/walk/service/WalkService.java
@@ -28,6 +28,7 @@ public class WalkService {
     public Walk createWalk(Long dogId, RequestWalkDto requestWalkDto){
         User user = userService.getAuthenticatedUser();
         Dog dog = dogService.findById(dogId);
+
         if(dog != null && dog.getUser().equals(user)) {
             Walk walk = requestWalkDto.toEntity(dog);
             return walkRepository.save(walk);
@@ -62,13 +63,11 @@ public class WalkService {
 
     public boolean deleteWalk(Long walkId){
         Walk walk = findById(walkId);
-        if(walk != null){
+        if(walk != null && dogService.validation(walk.getDog())){
             walkRepository.deleteById(walkId);
             return true;
         }else{
             return false;
         }
     }
-
-
 }
